Separate the mocker into its own file.
