+++ lib/unexpectedMitm.js
@@ -1,20 +1,21 @@
 /* global setImmediate, process, after, console */
 var messy = require('messy'),
-    createError = require('createerror'),
+    consumeReadableStream = require('./consumeReadableStream'),
     createMitm = require('mitm-papandreou'),
+    createSerializedRequestHandler = require('./createSerializedRequestHandler'),
     _ = require('underscore'),
     http = require('http'),
     https = require('https'),
+    formatHeaderObj = require('./formatHeaderObj'),
     fs = require('fs'),
     path = require('path'),
-    stream = require('stream'),
-    urlModule = require('url'),
     memoizeSync = require('memoizesync'),
     callsite = require('callsite'),
     detectIndent = require('detect-indent'),
     metadataPropertyNames = messy.HttpRequest.metadataPropertyNames.concat('rejectUnauthorized');
 
-var isNodeZeroTen = !!process.version.match(/v0.10/);
+var UnexpectedMitmMocker = require('./UnexpectedMitmMocker');
+
 // fallback to an inlined version of 0.12+ path.isAbsolute() for 0.10 compat
 var pathIsAbsolute = path.isAbsolute || function (path) {
     var len = path.length;
@@ -38,31 +39,12 @@ var pathIsAbsolute = path.isAbsolute || function (path) {
     return false;
 };
 
-// base error
-var UnexpectedMitmError = createError({ name: 'UnexpectedMitmError' });
-// marker errors
-var EarlyExitError = createError({ name: 'EarlyExitError' }, UnexpectedMitmError);
-var SawUnexpectedRequestsError = createError({ name: 'SawUnexpectedRequestsError' }, UnexpectedMitmError);
-var UnexercisedMocksError = createError({ name: 'UnexercisedMocksError' }, UnexpectedMitmError);
-
 function checkEnvFlag(varName) {
     return process.env[varName] === 'true';
 }
 
-function isRegExp(obj) {
-    return Object.prototype.toString.call(obj) === '[object RegExp]';
-}
-
 function isTextualBody(message, content) {
     return message.hasTextualContentType && bufferCanBeInterpretedAsUtf8(content);
-}
-
-function formatHeaderObj(headerObj) {
-    var result = {};
-    Object.keys(headerObj).forEach(function (headerName) {
-        result[messy.formatHeaderName(headerName)] = headerObj[headerName];
-    });
-    return result;
 }
 
 function bufferCanBeInterpretedAsUtf8(buffer) {
@@ -97,13 +79,6 @@ function trimHeaders(message) {
     delete message.headers['Transfer-Encoding'];
     delete message.headers.Connection;
     delete message.headers.Date;
-}
-
-function trimHeadersLower(message) {
-    delete message.headers.valuesByName['content-length'];
-    delete message.headers.valuesByName['transfer-encoding'];
-    delete message.headers.valuesByName.connection;
-    delete message.headers.valuesByName.date;
 }
 
 function trimMessage(message) {
@@ -153,219 +128,11 @@ function trimMessage(message) {
     return message;
 }
 
-function trimMockResponse(mockResponse) {
-    var responseProperties = {};
-    responseProperties.statusCode = mockResponse.statusCode;
-    responseProperties.statusMessage = mockResponse.statusMessage;
-    responseProperties.protocolName = mockResponse.protocolName;
-    responseProperties.protocolVersion = mockResponse.protocolVersion;
-    responseProperties.headers = formatHeaderObj(mockResponse.headers.valuesByName);
-
-    if (mockResponse.body) {
-        // read out the messy decoded body
-        responseProperties.body = mockResponse.body;
-    }
-
-    return responseProperties;
-}
-
 function trimRecordedExchange(recordedExchange) {
     return {
         request: trimMessage(recordedExchange.request),
         response: trimMessage(recordedExchange.response)
     };
-}
-
-function createSerializedRequestHandler(onRequest) {
-    var activeRequest = false,
-        requestQueue = [];
-
-    function processNextRequest() {
-        function cleanUpAndProceed() {
-            if (activeRequest) {
-                activeRequest = false;
-                setImmediate(processNextRequest);
-            }
-        }
-        while (requestQueue.length > 0 && !activeRequest) {
-            activeRequest = true;
-            var reqAndRes = requestQueue.shift(),
-                req = reqAndRes[0],
-                res = reqAndRes[1],
-                resEnd = res.end;
-            res.end = function () {
-                resEnd.apply(this, arguments);
-                cleanUpAndProceed();
-            };
-            // This happens upon an error, so we need to make sure that we catch that case also:
-            res.on('close', cleanUpAndProceed);
-            onRequest(req, res);
-        }
-    }
-
-    return function (req, res) {
-        requestQueue.push([req, res]);
-        processNextRequest();
-    };
-}
-
-function resolveExpectedRequestProperties(expectedRequestProperties) {
-    if (typeof expectedRequestProperties === 'string') {
-        expectedRequestProperties = { url: expectedRequestProperties };
-    } else if (expectedRequestProperties && typeof expectedRequestProperties === 'object') {
-        expectedRequestProperties = _.extend({}, expectedRequestProperties);
-    }
-    if (expectedRequestProperties) {
-        if (typeof expectedRequestProperties.url === 'string') {
-            var matchMethod = expectedRequestProperties.url.match(/^([A-Z]+) ([\s\S]*)$/);
-            if (matchMethod) {
-                expectedRequestProperties.method = expectedRequestProperties.method || matchMethod[1];
-                expectedRequestProperties.url = matchMethod[2];
-            }
-        }
-    } else {
-        expectedRequestProperties = {};
-    }
-    if (/^https?:\/\//.test(expectedRequestProperties.url)) {
-        var urlObj = urlModule.parse(expectedRequestProperties.url);
-        expectedRequestProperties.headers = expectedRequestProperties.headers || {};
-        if (Object.keys(expectedRequestProperties.headers).every(function (key) {
-            return key.toLowerCase() !== 'host';
-        })) {
-            expectedRequestProperties.headers.host = urlObj.host;
-        }
-        expectedRequestProperties.host = expectedRequestProperties.host || urlObj.hostname;
-        if (urlObj.port && typeof expectedRequestProperties.port === 'undefined') {
-            expectedRequestProperties.port = parseInt(urlObj.port, 10);
-        }
-
-        if (urlObj.protocol === 'https:' && typeof expectedRequestProperties.encrypted === 'undefined') {
-            expectedRequestProperties.encrypted = true;
-        }
-        expectedRequestProperties.url = urlObj.path;
-    }
-
-    var expectedRequestBody = expectedRequestProperties.body;
-    if (Array.isArray(expectedRequestBody) || (expectedRequestBody && typeof expectedRequestBody === 'object' && !isRegExp(expectedRequestBody) && (typeof Buffer === 'undefined' || !Buffer.isBuffer(expectedRequestBody)))) {
-        // in the case of a streamed request body and skip asserting the body
-        if (typeof expectedRequestBody.pipe === 'function') {
-            throw new Error('unexpected-mitm: a stream cannot be used to verify the request body, please specify the buffer instead.');
-        }
-        expectedRequestProperties.headers = expectedRequestProperties.headers || {};
-        if (Object.keys(expectedRequestProperties.headers).every(function (key) {
-            return key.toLowerCase() !== 'content-type';
-        })) {
-            expectedRequestProperties.headers['Content-Type'] = 'application/json';
-        }
-    }
-    return expectedRequestProperties;
-}
-
-function createMockResponse(responseProperties) {
-    var mockResponse = new messy.HttpResponse(responseProperties);
-    mockResponse.statusCode = mockResponse.statusCode || 200;
-    mockResponse.protocolName = mockResponse.protocolName || 'HTTP';
-    mockResponse.protocolVersion = mockResponse.protocolVersion || '1.1';
-    mockResponse.statusMessage = mockResponse.statusMessage || http.STATUS_CODES[mockResponse.statusCode];
-    return mockResponse;
-}
-
-function calculateBodyByteCount(chunk) {
-    var trailerIdx = findHeaderSeparator(chunk);
-    if (trailerIdx !== -1) {
-        return chunk.slice(trailerIdx + 4).length;
-    }
-
-    return 0;
-}
-
-function findHeaderSeparator(chunk) {
-    return chunk.toString().indexOf('\r\n\r\n');
-}
-
-function isTrailerChunk(chunk) {
-    return chunk.slice(-5).toString() === '0\r\n\r\n';
-}
-
-function attachConnectionHook(connection, callback) {
-    var bodyByteCount = 0;
-    var contentLengthByteCount = 0;
-    var hasContentLength = false;
-    var rawChunks = [];
-    var sawHeaders = false;
-    var _write = connection._write;
-
-    /*
-     * wrap low level write to gather response data and return raw data
-     * to callback when entire response is written
-     */
-    connection._write = function (chunk, encoding, cb) {
-        chunk = Buffer.isBuffer(chunk) ? chunk : new Buffer(chunk);
-
-        rawChunks.push(chunk);
-
-        if (!sawHeaders) {
-            var chunkString = chunk.toString();
-            var match;
-
-            match = chunkString.match(/^Transfer-Encoding: /m);
-            if (match) {
-                sawHeaders = true;
-            }
-
-            match = chunkString.match(/^Content-Length: (\d+)/m);
-            if (match) {
-                // handle the current chunk containing body bytes
-                bodyByteCount = calculateBodyByteCount(chunk);
-                contentLengthByteCount = +match[1];
-                hasContentLength = true;
-                sawHeaders = true;
-            }
-        } else {
-            bodyByteCount += chunk.length;
-        }
-
-        if ((isTrailerChunk(chunk) && !hasContentLength) ||
-                (hasContentLength && bodyByteCount === contentLengthByteCount)) {
-            /*
-             * explicitly execute the callback returning raw data here
-             * to ensure it is run before issuing the final write which
-             * will complete the request on the wire
-             */
-            callback(null, Buffer.concat(rawChunks));
-        }
-
-        _write.call(connection, chunk, encoding, cb);
-    };
-}
-
-function attachResponseHooks(res, callback) {
-    var hasEnded = false; // keep track of whether res.end() has been called
-
-    attachConnectionHook(res.connection, function (err, rawBuffer) {
-        if (hasEnded) {
-            callback(null, rawBuffer);
-        }
-    });
-
-    ['end', 'destroy'].forEach(function (methodName) {
-        var orig = res[methodName];
-        res[methodName] = function (chunk, encoding) {
-            if (methodName === 'end') {
-                hasEnded = true;
-            }
-
-            var returnValue = orig.apply(this, arguments);
-
-            if (methodName === 'destroy') {
-                hasEnded = true;
-                callback(null);
-            }
-
-            // unconditionally return value given we do not wrap write
-            return returnValue;
-        };
-    });
 }
 
 function determineCallsite(stack) {
@@ -543,60 +310,6 @@ module.exports = {
             } else {
                 console.warn('unexpected-mitm: Could not find the right place to inject the recorded exchanges into ' + sourceFileName + ' (around line ' + sourceLineNumber + '): ' + stringify(recordedExchanges, indentationWidth, expect));
             }
-        }
-
-        function consumeReadableStream(readableStream, options) {
-            options = options || {};
-            var skipConcat = !!options.skipConcat;
-
-            return expect.promise(function (resolve, reject) {
-                var chunks = [];
-                readableStream.on('data', function (chunk) {
-                    chunks.push(chunk);
-                }).on('end', function (chunk) {
-                    resolve({ body: skipConcat ? chunks : Buffer.concat(chunks) });
-                }).on('error', function (err) {
-                    resolve({ body: skipConcat ? chunks : Buffer.concat(chunks), error: err });
-                });
-            });
-        }
-
-        function getMockResponse(responseProperties) {
-            var mockResponse;
-            var mockResponseError;
-            if (Object.prototype.toString.call(responseProperties) === '[object Error]') {
-                mockResponseError = responseProperties;
-            } else {
-                mockResponse = createMockResponse(responseProperties);
-            }
-
-            return expect.promise(function () {
-                if (!mockResponseError && mockResponse && mockResponse.body && typeof mockResponse.body.pipe === 'function') {
-                    return consumeReadableStream(mockResponse.body).then(function (result) {
-                        if (result.error) {
-                            mockResponseError = result.error;
-                        }
-                        if (result.body) {
-                            mockResponse.unchunkedBody = result.body;
-                        }
-                    });
-                }
-            }).then(function () {
-                if (mockResponse && !mockResponseError && (Array.isArray(mockResponse.body) || (mockResponse.body && typeof mockResponse.body === 'object' && (typeof Buffer === 'undefined' || !Buffer.isBuffer(mockResponse.body))))) {
-                    if (!mockResponse.headers.has('Content-Type')) {
-                        mockResponse.headers.set('Content-Type', 'application/json');
-                    }
-                }
-                return [ mockResponse, mockResponseError ];
-            });
-        }
-
-        function observeResponse(res) {
-            return expect.promise(function (resolve, reject) {
-                attachResponseHooks(res, function (err, rawBuffer) {
-                    resolve(rawBuffer);
-                });
-            });
         }
 
         function performRequest(requestResult) {
@@ -889,11 +602,14 @@ module.exports = {
                     return verifyBlock;
                 });
 
-                var assertionPromise = modules.Mocker({
-                    descriptions: requestDescriptions,
-                    consume: function () {
-                        return expect.shift();
-                    }
+                var mocker = new UnexpectedMitmMocker({
+                    descriptions: requestDescriptions
+                });
+
+                var assertionPromise = mocker.mock(function () {
+                    return expect.shift();
+                }).then(function (result) {
+                    return [result.timeline, result.fulfilmentValue];
                 }).spread(function (timeline, fulfilmentValue) {
                     var lastEventOrError = null;
 
@@ -970,309 +686,5 @@ module.exports = {
 
                 return assertionPromise;
             });
-
-        var modules = {
-            Mocker: function (options) {
-                options = options || {};
-                var requestDescriptions = options.descriptions;
-                var consumptionPromise = options.consume;
-
-                var mitm = createMitm();
-
-                // Keep track of the current requestDescription
-                var nextRequestDescriptionIndex = 0;
-
-                // Keep track of the http/https agents that we have seen
-                // during the test so we can clean up afterwards:
-                var seenAgents = [];
-
-                // accumulator for events
-                var timeline = [];
-
-                return expect.promise(function (resolve, reject) {
-                    mitm.on('request', createSerializedRequestHandler(function (req, res) {
-                        if (!res.connection) {
-                            // I've observed some cases where keep-alive is
-                            // being used and we end up with an extra "phantom
-                            // request event" even though only requeest is being
-                            // issued. Seems like a bug in mitm.
-                            // It goes without saying that we should try to get
-                            // rid of this. Hopefully something will happen
-                            // on https://github.com/moll/node-mitm/pull/36
-                            return;
-                        }
-                        var clientSocket = req.connection._mitm.client;
-                        var clientSocketOptions = req.connection._mitm.opts;
-                        if (typeof clientSocketOptions.port === 'string') {
-                            // The port could have been defined as a string in a 3rdparty library doing the http(s) call, and that seems to be valid use of the http(s) module
-                            clientSocketOptions = _.defaults({
-                                port: parseInt(clientSocketOptions.port, 10)
-                            }, clientSocketOptions);
-                        }
-                        var agent = clientSocketOptions.agent || (res.connection.encrypted ? https : http).globalAgent;
-                        if (seenAgents.indexOf(agent) === -1) {
-                            seenAgents.push(agent);
-                        }
-                        var currentDescription = requestDescriptions[nextRequestDescriptionIndex];
-                        nextRequestDescriptionIndex += 1;
-                        var hasRequestDescription = !!currentDescription,
-                            metadata =
-                                _.defaults(
-                                    { encrypted: Boolean(res.connection.encrypted) },
-                                    _.pick(clientSocketOptions, messy.HttpRequest.metadataPropertyNames),
-                                    _.pick(clientSocketOptions && clientSocketOptions.agent && clientSocketOptions.agent.options, messy.HttpRequest.metadataPropertyNames)
-                                ),
-                            requestDescription = currentDescription,
-                            requestProperties,
-                            responseProperties = requestDescription && requestDescription.response,
-                            expectedRequestProperties;
-
-                        expect.promise(function () {
-                            expectedRequestProperties = resolveExpectedRequestProperties(requestDescription && requestDescription.request);
-                        }).then(function () {
-                            return consumeReadableStream(req, {skipConcat: true});
-                        }).then(function (result) {
-                            if (result.error) {
-                                // TODO: Consider adding support for recording this (the request erroring out while we're consuming it)
-                                throw result.error;
-                            }
-                            requestProperties = _.extend({
-                                method: req.method,
-                                path: req.url,
-                                protocolName: 'HTTP',
-                                protocolVersion: req.httpVersion,
-                                headers: req.headers,
-                                unchunkedBody: Buffer.concat(result.body)
-                            }, metadata);
-
-                            if (!hasRequestDescription) {
-                                // there was no mock so arrange "<no response>"
-                                assertMockResponse(null);
-
-                                /*
-                                 * We wish to cause the generation of a diff
-                                 * so we arrange to enter our 'success' path
-                                 * but ensure the delegated assertion fully
-                                 * completes by signalling an error condition.
-                                 *
-                                 * Note the use of the single reject/resolve
-                                 * behaviour of promises: while the delegated
-                                 * assertion is reject()ed, we have already
-                                 * resolve()d thus the rejection of the former
-                                 * is effectively ignored and we proceed with
-                                 * our output.
-                                 */
-
-                                // cancel the delegated assertion
-                                throw new SawUnexpectedRequestsError('unexpected-mitm: Saw unexpected requests.');
-                            }
-
-                            if (typeof responseProperties === 'function') {
-                                // reset the readable req stream state
-                                stream.Readable.call(req);
-
-                                // read stream data from the buffered chunks
-                                req._read = function () {
-                                    this.push(result.body.shift() || null);
-                                };
-
-                                if (isNodeZeroTen) {
-                                    /*
-                                     * As is mentioned in the streams documentation this
-                                     * call can be issued to kick some of the machinery
-                                     * and is apparently done within the standard lib.
-                                     */
-                                    req.read(0);
-                                }
-
-                                // call response function inside a promise to catch exceptions
-                                return expect.promise(function () {
-                                    responseProperties(req, res);
-                                }).then(function () {
-                                    return [null, null];
-                                }).caught(function (err) {
-                                    // ensure delivery of the caught error to the underlying socket
-                                    return [null, err];
-                                });
-                            } else {
-                                return getMockResponse(responseProperties);
-                            }
-                        }).spread(function (mockResponse, mockResponseError) {
-                            if (!(mockResponse || mockResponseError)) {
-                                return;
-                            }
-
-                            var originalErrorMode = expect.errorMode;
-                            expect.errorMode = 'default';
-                            return expect.promise(function () {
-                                var assertionMockRequest = new messy.HttpRequest(requestProperties);
-                                trimHeadersLower(assertionMockRequest);
-                                return expect(assertionMockRequest, 'to satisfy', expectedRequestProperties);
-                            }).then(function () {
-                                expect.errorMode = originalErrorMode;
-                                // continue thus respond
-                                deliverMockResponse(mockResponse, mockResponseError);
-                            }).catch(function (e) {
-                                assertMockResponse(mockResponse, mockResponseError);
-                                throw new EarlyExitError('Seen request did not match the expected request.');
-                            });
-                        }).caught(function (e) {
-                            /*
-                             * Given an error occurs, the deferred assertion
-                             * will still be pending and must be completed. We
-                             * do this by signalling the error on the socket.
-                             */
-
-                            // cancel the delegated assertion
-                            try {
-                                clientSocket.emit('error', e);
-                            } finally {
-                                /*
-                                 * If an something was thrown trying to signal
-                                 * an error we have little choice but to try to
-                                 * reject the assertion. This is only safe with
-                                 * Unexpected 10.15.x and above.
-                                 */
-
-                                timeline.push(e);
-                                reject(e);
-                            }
-                        });
-
-                        function assertMockResponse(mockResponse, mockResponseError) {
-                            var mockRequest = new messy.HttpRequest(requestProperties);
-
-                            var httpExchange = new messy.HttpExchange({request: mockRequest});
-                            if (mockResponse) {
-                                httpExchange.response = mockResponse;
-                            } else if (mockResponseError) {
-                                httpExchange.response = mockResponseError;
-                            }
-
-                            var eventSpec = null;
-                            // only attempt request validation with a mock
-                            if (hasRequestDescription) {
-                                eventSpec = {request: expectedRequestProperties};
-                            }
-
-                            /*
-                             * We explicitly do not complete the promise
-                             * when the last request for which we have a
-                             * mock is seen.
-                             *
-                             * Instead simply record any exchanges that
-                             * passed through and defer the resolution or
-                             * rejection to the handler functions attached
-                             * to the delegated assertion execution below.
-                             */
-
-                            timeline.push({
-                                exchange: httpExchange,
-                                spec: eventSpec
-                            });
-                        }
-
-                        function deliverMockResponse(mockResponse, mockResponseError) {
-                            setImmediate(function () {
-                                var nonEmptyMockResponse = false;
-                                if (mockResponse) {
-                                    res.statusCode = mockResponse.statusCode;
-                                    mockResponse.headers.getNames().forEach(function (headerName) {
-                                        mockResponse.headers.getAll(headerName).forEach(function (value) {
-                                            nonEmptyMockResponse = true;
-                                            res.setHeader(headerName, value);
-                                        });
-                                    });
-                                    var unchunkedBody = mockResponse.unchunkedBody;
-                                    if (typeof unchunkedBody !== 'undefined' && unchunkedBody.length > 0) {
-                                        nonEmptyMockResponse = true;
-                                        res.write(unchunkedBody);
-                                    } else if (nonEmptyMockResponse) {
-                                        res.writeHead(res.statusCode || 200);
-                                    }
-                                }
-                                if (mockResponseError) {
-                                    setImmediate(function () {
-                                        clientSocket.emit('error', mockResponseError);
-                                        assertMockResponse(mockResponse, mockResponseError);
-                                    });
-                                } else {
-                                    res.end();
-                                }
-                            });
-                        }
-
-                        /*
-                         * hook final write callback to record raw data and immediately
-                         * assert request so it occurs prior to request completion
-                         */
-                        observeResponse(res).then(function (rawBuffer) {
-                            var mockResponse = createMockResponse(rawBuffer);
-                            assertMockResponse(mockResponse);
-                        });
-                    }));
-
-                    expect.promise(function () {
-                        return consumptionPromise();
-                    }).then(function (fulfilmentValue) {
-                        /*
-                         * Handle the case of specified but unprocessed mocks.
-                         * If there were none remaining immediately complete.
-                         *
-                         * Where the driving assertion resolves we must check
-                         * if any mocks still exist. If so, we add them to the
-                         * set of expected exchanges and resolve the promise.
-                         */
-                        var hasRemainingRequestDescriptions = nextRequestDescriptionIndex < requestDescriptions.length;
-                        if (hasRemainingRequestDescriptions) {
-                            // exhaust remaining mocks using a promises chain
-                            return (function nextItem() {
-                                var remainingDescription = requestDescriptions[nextRequestDescriptionIndex];
-                                nextRequestDescriptionIndex += 1;
-                                if (remainingDescription) {
-                                    var expectedRequestProperties;
-
-                                    return expect.promise(function () {
-                                        expectedRequestProperties = resolveExpectedRequestProperties(remainingDescription.request);
-                                    }).then(function () {
-                                        return getMockResponse(remainingDescription.response);
-                                    }).spread(function (mockResponse, mockResponseError) {
-                                        var spec = {
-                                            request: expectedRequestProperties,
-                                            response: mockResponseError || trimMockResponse(mockResponse)
-                                        };
-
-                                        timeline.push({ spec: spec });
-
-                                        return nextItem();
-                                    });
-                                } else {
-                                    throw new UnexercisedMocksError();
-                                }
-                            })();
-                        } else {
-                            resolve([fulfilmentValue]);
-                        }
-                    }).caught(function (e) {
-                        timeline.push(e);
-                        reject(e);
-                    });
-                }).spread(function (fulfilmentValue) {
-                    return [timeline, fulfilmentValue];
-                }).catch(function () {
-                    // given we will later fail on "to satisfy" pass a null fulfilment value
-                    return [timeline, null];
-                }).finally(function () {
-                    seenAgents.forEach(function (agent) {
-                        if (agent.freeSockets) {
-                            Object.keys(agent.freeSockets).forEach(function (key) {
-                                agent.freeSockets[key] = [];
-                            });
-                        }
-                    });
-                    mitm.disable();
-                });
-            }
-        };
     }
 };
